[#34] feat : 게시글 조회, 작성 시 각 해시태그 갱신 확인

+++ src/main/java/com/sparta/elevenbookshelf/entity/Hashtag.java
@@ -55,6 +55,6 @@ public class Hashtag {
     }
 
     public void updateTier() {
-        this.tier = this.count/10;
+        this.tier = (double) this.count /10;
     }
 }
